Migrate code base from `geoenvo` to `geoenv`
Migrate the code base from `geoenvo` (https://github.com/clnsmth/geoenvo)
to `geoenv` (here). The package was originally developed under the name
`geoenvo` but was renamed to avoid potential naming conflicts and to
reflect the broader scope of the package, which is intended to support
various semantic resources beyond the Environment Ontology (ENVO).